Moved FastBoot service to app and refactored the code.
* Moved Shoebox and request object to separate utils module
* Moving the chagnes to app folder allows overriding of request-object and shoebox with app specific custom objects
* Added test for request object
